Move http middleware to rpc lib
